Refactor astForVariableIdentifier
