+++ joern-cli/frontends/rubysrc2cpg/src/main/antlr4/io/joern/rubysrc2cpg/parser/RubyParser.g4
@@ -484,8 +484,8 @@ jumpExpression
 // --------------------------------------------------------
 
 variableReference
-    :   variableIdentifier
-    |   pseudoVariableIdentifier
+    :   variableIdentifier                                                                                          # variableIdentifierVariableReference
+    |   pseudoVariableIdentifier                                                                                    # pseudoVariableIdentifierVariableReference
     ;
 
 variableIdentifier
