+++ joern-cli/frontends/rubysrc2cpg/src/test/scala/io/joern/rubysrc2cpg/parser/InvocationWithParenthesesTests.scala
@@ -68,7 +68,7 @@ class InvocationWithParenthesesTests extends RubyParserAbstractTest {
             |    Association
             |     PrimaryExpression
             |      VariableReferencePrimary
-            |       VariableReference
+            |       VariableIdentifierVariableReference
             |        VariableIdentifier
             |         region
             |     :
